Add async wrapping for completion WebSocket (#102)
**Before**:

Whenever we were running llama.cpp code, we were actually blocking the
main event loop every time we iterated over the completion
chunks...:facepalm: This would cause all other request processing to
completely halt while any completion requests were outstanding.

The main problem was that we were running this directly on the main
event loop thread, but even moving this to a background thread isn't
enough because of the GIL, so we actually need to move processing to a
separate multiprocessing process. We also need to be careful that in the
main event loop that passes data to the background process doesn't
accidentally cause blocking behavior with the queue.

**After**:

We have moved llama.cpp completion to a background thread that is spun
up on-demand and integrated with the async WebSocket interface.

This has alleviated the issue where completions block processing of any
other endpoint logic.

Additionally, I've added an endpoint at `/admin/stack` that dumps a
stacktrace of all Python threads as a debugging tool for us to use to
inspect the process going forward.